feat: Ability to specify "tag_format" for pre-release action.
In some cases we may need to set [tag_format](https://semantic-release.gitbook.io/semantic-release/usage/configuration#tagformat) for the "semantic-release", for example to have an app name as a tag prefix for the helm chart.
Although `tag_format` may be defined in the config, it does not work very well for the monorepo, when multiple configs may be needed for multiple apps, which has some issues - [ref](https://github.com/semantic-release/semantic-release/issues/1592#issuecomment-1695740555). Moreover, the underlying action we use always tries to download the config from NPM instead of using a local file.